refactor: 공통 api, type 상위 폴더로 이동
