+++ apps/web/app/[locale]/(with-layout)/brand/applicants/apis/index.ts
@@ -1,10 +1,6 @@
 import { apiRequest } from 'app/api/apiRequest';
 
-import {
-  ApplicantsApiResponse,
-  BrandCampaignInfosApiResponse,
-  GetApplicantsParams,
-} from '../types';
+import { ApplicantsApiResponse, GetApplicantsParams } from '../types';
 
 /**
  * 브랜드 지원자 목록 조회 API
@@ -30,20 +26,6 @@ export const getApplicants = async (
   });
   return response;
 };
-
-/**
- * 브랜드 캠페인 정보 목록 조회 API
- * @returns 브랜드의 모든 캠페인 정보 목록
- */
-
-export const getBrandCampaignInfos =
-  async (): Promise<BrandCampaignInfosApiResponse> => {
-    const response = await apiRequest<BrandCampaignInfosApiResponse>({
-      endPoint: '/api/brands/my/campaigns/infos',
-      method: 'GET',
-    });
-    return response;
-  };
 
 export const approveApplicants = async (
   campaignId: number,
