+++ apps/web/app/[locale]/(with-layout)/brand/applicants/types/index.ts
@@ -38,19 +38,3 @@ export interface ApplicantsApiResponse {
     };
   };
 }
-
-export interface CampaignInfo {
-  campaignId: number;
-  campaignTitle: string;
-  startDate: string;
-  endDate: string;
-}
-
-export interface BrandCampaignInfosApiResponse {
-  success: boolean;
-  status: number;
-  message: string;
-  data: {
-    campaignInfos: CampaignInfo[];
-  };
-}
