+++ apps/web/app/[locale]/(with-layout)/brand/types/index.ts
@@ -1,0 +1,15 @@
+export interface CampaignInfo {
+  campaignId: number;
+  campaignTitle: string;
+  startDate: string;
+  endDate: string;
+}
+
+export interface BrandCampaignInfosApiResponse {
+  success: boolean;
+  status: number;
+  message: string;
+  data: {
+    campaignInfos: CampaignInfo[];
+  };
+}
